query_alert/requirements.txt
pandas
pandas-gbq
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-api-python-client